web/requirements.txt
cbor2==4.1.2
cryptography==3.3.2
flask==1.1.2
flask_compress==1.9.0
flask_session==0.3.1
gevent==21.1.2
gnupg==2.3.1
libtiff==0.4.2
numpy==1.16.6
pathlib==1.0.1
pdf2image==1.13.1
pillow==6.2.2
psycopg2-binary==2.8.6
pyBarcode==0.8b1
pycrypto==2.6.1
pyotp==2.3.0
pytz==2021.1
pyzbar==0.1.8
qrcode==6.1
redis==3.5.3
webauthn==0.4.7
